requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Django==5.2.8
django-mama-cas==2.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fastapi==0.123.9
Flask==2.3.3
Flask-CAS==1.0.2
flask-cors==6.0.1
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-multipart==0.0.20
requests==2.32.5
sqlparse==0.5.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
Werkzeug==3.1.4
xmltodict==1.0.2
